pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cv-agent"
version = "0.1.0"
description = "Executive CV tailoring + semi-auto application submission pipeline"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.40",
    "beautifulsoup4>=4.12",
    "trafilatura>=1.12",
    "requests>=2.32",
    "tenacity>=9.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "feedparser>=6.0",
    "playwright>=1.47",
    "weasyprint>=62.0",
    "markdown-it-py>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.6",
]

[project.scripts]
cv-search = "cv_agent.cli:search_main"
cv-generate = "cv_agent.cli:generate_main"
cv-submit = "cv_agent.cli:submit_main"
cv-reject = "cv_agent.cli:reject_main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"cv_agent.render" = ["templates/*.html"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
